fix: pass base to URL only when value is provided (#32)
On Safari attempting to pass `base` as `undefined` results with `TypeError` as seen on https://responsive-ng-router.netlify.app.
Based on type definition for URL, `base` property is optional and of type `string | URL`. While most browsers consider passing undefined and omitting the property same, Safari is somewhat strict in this matter and throws an exception.